agent/rag/requirements.txt
llama-index-core==0.14.3
llama-index-embeddings-ollama==0.8.3
llama-index-llms-ollama==0.7.4
pymupdf
llama-index-readers-file
